test: move Ruby specific tests to its own file for ownership reasons
